Renamed mixin package to mixins for consistency
