+++ src/main/resources/mcct.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "fr.smyler.mcct.mixin",
+  "package": "fr.smyler.mcct.mixins",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
   ],
